+++ the_project/constants.py
@@ -1,0 +1,15 @@
+# Layer Names
+LAYER_NAME_BACKGROUND = "background"
+LAYER_NAME_FOREGROUND = "foreground"
+LAYER_NAME_OBJECTS = "game_objects"
+
+# Asset Directions
+TEXTURE_MAIN_PATH = f"assets/maps/map_assets/"
+TEXTURE_DICT_PATH = {
+    "turret": "building/turret/turret",
+    "base": "building/base/base",
+    "player": "non_building/player/player"
+}
+
+
+